ci: macos

+++ libbsdfetch/http.c
@@ -462,7 +462,7 @@ http_cmd(conn_t *conn, const char *fmt, ...)
 		return (-1);
 	}
 
-	ED_TRACE(("cmd: <%.*s> (%d)\n", (int)(len - 2), msg, len))
+	ED_TRACE(("cmd: <%.*s> (%u)\n", (int)(len - 2), msg, (unsigned)len))
 	r = fetch_write(conn, msg, len);
 	free(msg);
